Bug Fix test_my_prompt_custom_script.py
Discovered a `job_no` calculation error when `excluded_keyword` was used. I think this should be resolved by keeping track of the number of times a job was skipped due to an excluded keyword and subtracting the number from the `job_no` update calculation. I think this should work because jobs that are skipped due to excluded keywords should not be counted. This bug would sometimes cause the number of jobs `job_no` to be larger than the total jobs `job_count`.